extract CardViewHolder from Card model
